Replaced "param()" with "param ()" for consistency.
Most examples in this project use the `param` keywork with a space
before the parentheses, like `param (` or `param ()`. However,
some examples omit the space, or capitalize the P, like `param(` or
`Param(`. This is a needless inconsistency, which this PR addresses.
Note that two instances of `param()` are left in tact, since they are
in a TODO file.